% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sf.cartogramR.R
\name{as.sf.cartogramR}
\alias{as.sf.cartogramR}
\title{Coerce a cartogramR to a sf object}
\usage{
\method{as.sf}{cartogramR}(x, ...)
}
\arguments{
\item{x}{a cartogramR object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
a sf object including all the data (attributes) contained
in the original sf object used to construct the cartogram and
\itemize{
\item original areas of region (\code{orig_area})
\item final/deformed areas of region (\code{final_area})
\item target areas of region (\code{target_area})
\item original centers (\code{x_orig_centers} and \code{y_orig_centers})
\item final centers (\code{x_final_centers} and \code{y_final_centers})
}
}
\description{
Coerce a cartogramR to a sf object returning the sf object used to
construct the cartogram with the cartogram as geometry and some
more attributes
}
