% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sf.R
\name{as.sf}
\alias{as.sf}
\title{Coerce an object to a sf object}
\usage{
as.sf(x, ...)
}
\arguments{
\item{x}{object to be coerced}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
an sf object
}
\description{
Coerce an object to a sf object
}
