% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sfc.cartogramR.R
\name{as.sfc.cartogramR}
\alias{as.sfc.cartogramR}
\title{Coerce a cartogramR to a sfc object}
\usage{
\method{as.sfc}{cartogramR}(x, ...)
}
\arguments{
\item{x}{a cartogramR object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
a sfc object
}
\description{
Coerce a cartogramR to a  sfc object extracting the component cartogram of
the cartogramR object
}
