% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sfc.R
\name{as.sfc}
\alias{as.sfc}
\title{Coerce an object to a sfc object}
\usage{
as.sfc(x, ...)
}
\arguments{
\item{x}{object to be coerced}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
a sfc object
}
\description{
Coerce an object to a sfc object
}
