% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sfmultipolygon.R
\name{as.sfmultipolygon}
\alias{as.sfmultipolygon}
\title{Transform a sf object with several rows (polygons) by region to an
sf object with one row by region and thus one multipolygon by region}
\usage{
as.sfmultipolygon(data, idregion, closepolygon = FALSE)
}
\arguments{
\item{data}{a sf object}

\item{idregion}{a character string which indicates the name of the
column (in \code{data} object) which contains the region
identifier.}

\item{closepolygon}{a boolean (default to \code{FALSE}) if \code{TRUE} it
controls if polygons are closed and if not add the first
vertice at the end.}
}
\value{
a sf object with one row by region and one multipolygon by
region.
}
\description{
Transform a sf object with several rows (polygons) by region to an
sf object with one row by region and thus one multipolygon by region
}
