% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogramR-deprecated.R, R/geom_cartogramR.R
\name{cartogramR-deprecated}
\alias{cartogramR-deprecated}
\alias{geom_cartogramR}
\title{Deprecated functions in package \pkg{cartogramR}.}
\usage{
geom_cartogramR(sfgeom, carto, verbose = FALSE)
}
\arguments{
\item{sfgeom}{a sf or a sfc object which contains simple feature geometry of
types in the following `POINT`, `MULTIPOINT`, `LINESTRING`,
`MULTILINESTRING`, `POLYGON`, `MULTIPOLYGON}

\item{carto}{a cartogramR object}

\item{verbose}{a boolean object to set on verbose mode (default to `FALSE`)}
}
\value{
a sf or a sfc object which contains simple feature geometry transformed
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.

Please use [cartogramR::warp_features()] instead.
}
\details{
The following functions are deprecated and will be made defunct; use
the replacement indicated below:
\itemize{
\item \verb{deprecated-function}: new-function
\item \code{geom_cartogramR}: \link{warp_features}
}
}
\examples{
\donttest{
  data(usa)
  carto <- cartogramR(usa, "electors64")
  LA <- sf::st_sfc(sf::st_point(c(-118.243685, 34.052234)))
  sf::st_crs(LA) <- 4326
  moregeom <- geom_cartogramR(LA, carto)
  plot(carto)
  plot(moregeom, add=TRUE, col=2, pch=15)
}

}
\keyword{internal}
