% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_between_vertices.R
\name{dist_between_vertices}
\alias{dist_between_vertices}
\title{Analyse some of the grid options}
\usage{
dist_between_vertices(data)
}
\arguments{
\item{data}{a sf object to be used in cartogram.}
}
\value{
a `dbv.cartogramR` object which is a data-table 
        which contains distance between vertices (`dbv`) and polygons names
        (`L1`, `L2`, `L3`) inherited from [sf::st_coordinates]
}
\description{
Analyse some of the grid options
}
\examples{
  data(usa)
  dbv <- dist_between_vertices(data=usa)
  summary(dbv)
}
