\name{france_rivers}
\docType{data}
\alias{france_rivers}
\title{Map of the population of mainland France (year 2020)}
\description{
  This data set is a basemap of mainland french rivers (\code{name} or \code{id}).
}
\usage{data(france_rivers)}
\source{\url{https://geoservices.ign.fr/documentation/diffusion/telechargement-donnees-libres.html}}
\format{A sf object containing 2 columns of data and the geometry
        Projected CRS: RGF93 / Lambert-93 (EPSG : 2154)}
\keyword{datasets}
