% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cartogramR.R
\name{plot.cartogramR}
\alias{plot.cartogramR}
\title{Plot a cartogram object}
\usage{
\method{plot}{cartogramR}(x, ...)
}
\arguments{
\item{x}{a cartogram object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Plot a cartogram object
}
\examples{
\donttest{
  data(usa)
  carto <- cartogramR(usa, "electors64")
  plot(carto)
}

}
