% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cartogramR.R
\name{print.cartogramR}
\alias{print.cartogramR}
\title{Print a cartogram object}
\usage{
\method{print}{cartogramR}(x, ...)
}
\arguments{
\item{x}{a cartogramR object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
No return value, called for side effects
}
\description{
Print a cartogram object
}
