% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.cartogramR.R
\name{print.summary.cartogramR}
\alias{print.summary.cartogramR}
\title{Print a summary of a cartogram object}
\usage{
\method{print}{summary.cartogramR}(x, ...)
}
\arguments{
\item{x}{a summary.cartogramR object}

\item{\\dots}{arguments passed to or from other methods. The following
argument is available at this level :  digits, the number of significant
digits to use when printing.}
}
\value{
x.
}
\description{
Print a summary of a cartogram object
}
\examples{
\donttest{
  data(usa)
  carto <- cartogramR(usa, "electors64")
  summary(carto)
}

}
