% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dbv.cartogramR.R
\name{summary.dbv.cartogramR}
\alias{summary.dbv.cartogramR}
\title{Summary of a dbv.cartogram object}
\usage{
\method{summary}{dbv.cartogramR}(object, ...)
}
\arguments{
\item{object}{a dbv.cartogramR object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
a data-table which contains by region (\code{L3})
\itemize{
\item the sample quantiles corresponding to the probability
0.8, 0.85, ...,1
\item the total number of vertices divided by the
perimeter of the region (the sum of all polygons
perimeter of the region, \code{NbyPerim})
}
}
\description{
Summary of a dbv.cartogram object
}
\examples{
  data(usa)
  dbv <- dist_between_vertices(data=usa)
  summary(dbv)

}
