% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gridanalysis.cartogramR.R
\name{summary.gridanalysis.cartogramR}
\alias{summary.gridanalysis.cartogramR}
\title{Summary of a gridanalysis.cartogram object}
\usage{
\method{summary}{gridanalysis.cartogramR}(object, ...)
}
\arguments{
\item{object}{a gridanalysis.cartogramR object}

\item{\\dots}{arguments passed to or from other methods.}
}
\value{
A vector which indicate the grid size necessary to have more than \code{steps}
grid points in each polygon
}
\description{
Summary of a gridanalysis.cartogram object
}
\examples{
\donttest{
  data(usa)
  ga <- grid_analysis(data=usa, gridpower2=4:9)
  summary(ga)
}
}
