\name{usa}
\docType{data}
\alias{usa}
\title{Map of the number of electors in each state of the USA}
\description{
  This data set is a basemap of the conterminous USA with the the number of electors in the 49 states from 1964 to 2020.
}
\usage{data(usa)}
\format{A sf object containing 24 columns of data and the geometry. Projected CRS: US National Atlas Equal Area (EPSG:2163)}
\keyword{datasets}
