% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogramR-deprecated.R
\name{cartogramR-deprecated}
\alias{cartogramR-deprecated}
\title{Deprecated functions in package \pkg{cartogramR}.}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\details{
The following functions are deprecated and will be made defunct; use
the replacement indicated below:
\itemize{
\item \verb{deprecated-function}: new-function
\item \code{geom_cartogramR}: \link{warp_features}
}
}
\keyword{internal}
