\name{france_rivers}
\docType{data}
\alias{france_rivers}
\title{Map of french rivers  of mainland France}
\description{
  This data set is a basemap of mainland french rivers (\code{name} or \code{id}).
}
\usage{data(france_rivers)}
\source{\url{https://geoservices.ign.fr/telechargement-api}}
\format{A sf object containing 2 columns of data and the geometry
        Projected CRS: RGF93 / Lambert-93 (EPSG : 2154)}
\keyword{datasets}
